Enhance production solver and validation scripts
- Set default PyTorch dtype to float32 to avoid half precision issues.
- Improved checkpoint fix application by importing the module properly and applying attention mask fixes.
- Updated quick validation report with new memory and disk usage metrics.
- Adjusted precision settings in scalability optimizer to use float32.
- Disabled fp16 in comprehensive training configuration due to compatibility issues.
- Added a cleanup script to remove cache files and temporary directories.
- Created an expanded data generator for minimum production requirements.
- Implemented a quick checkpoint test to validate checkpoint fixes with minimal data.
- Developed an ultra-fast validation script to test critical components and training efficiency.

+++ src/m3tm/search/multimodal_fusion.py
@@ -815,7 +815,7 @@ class ScalabilityOptimizer:
         estimated_memory = self._estimate_memory_usage(dataset_size)
         if estimated_memory > self.max_memory_gb:
             optimization_config["use_quantization"] = True
-            optimization_config["precision"] = "fp16"
+            optimization_config["precision"] = "fp32"  # Use fp32 instead of fp16 to avoid Half precision issues
             optimization_config["chunk_size"] = max(1000, optimization_config["chunk_size"] // 2)
             
         self.optimization_stats["dataset_optimizations"].append({
